requirements.txt
asgiref==3.5.2
Babel==2.11.0
Django==4.1.2
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
phonenumbers==8.13.2
pytz==2022.6
sqlparse==0.4.3
